# ⚡️ Speed up function `zoom_image` by 56%