--- a/unstructured_inference/models/tables.py
+++ b/unstructured_inference/models/tables.py
@@ -766,16 +766,17 @@ def zoom_image(image: PILImage.Image, zoom: float) -> PILImage.Image:
     if zoom <= 0:
         # no zoom but still does dilation and erosion
         zoom = 1
+    
+    kernel = np.ones((1, 1), np.uint8)
     new_image = cv2.resize(
-        cv2.cvtColor(np.array(image), cv2.COLOR_RGB2BGR),
+        cv2.cvtColor(np.asarray(image), cv2.COLOR_RGB2BGR),
         None,
         fx=zoom,
         fy=zoom,
         interpolation=cv2.INTER_CUBIC,
     )
 
-    kernel = np.ones((1, 1), np.uint8)
-    new_image = cv2.dilate(new_image, kernel, iterations=1)
-    new_image = cv2.erode(new_image, kernel, iterations=1)
+    new_image = cv2.dilate(new_image, kernel, iterations=1, dst=new_image)
+    new_image = cv2.erode(new_image, kernel, iterations=1, dst=new_image)
 
     return PILImage.fromarray(new_image)